refactor!: remove the libuild root export - it's a CLI, not a library
build()/publish() were never a designed API: the "." export existed
because zero-config discovery promoted the top-level implementation
file. src/libuild.ts moves to src/internal/libuild.ts alongside the
other internals, and the package surface is exactly the bin plus
@b9g/libuild/test. Importing "@b9g/libuild" now fails with
ERR_PACKAGE_PATH_NOT_EXPORTED (verified).

detectMainEntry rule 7 dies with it: with nothing declared (no
exports["."], main, module, index entry, multiple entries, no name
match) the build emits NO root export instead of alphabetically
electing one - the old fallback would have silently made "." point at
cli.js here. Single-entry (rule 5) and name-match (rule 6, which crank
relies on) are unchanged; mainEntry-undefined was already a supported
state downstream ("skip for bin-only packages").

Also: the 5s-default timeout immunization extends to every
fixture-building test file (127 tests across 12 files) - underload
this suite run took 190s vs the usual 138, and the un-timeouted files
were failing one by one with killed npm/build children.

Verified: dist exports = ./cli, ./cli.js, ./test, ./test.js,
./package.json; consumers green on bun/node/chromium; 206/0.
