+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [0.2.20] - 2026-08-18
 
+### Removed
+- **The `.` / `./libuild` root export is gone: libuild is a CLI, not a library.** `build()`/`publish()` were never a designed API - the root export existed because zero-config discovery promoted a top-level implementation file. The implementation now lives in `src/internal/` (with the other internals: esbuild recovery, snapshots, the test runners) and the package surface is exactly the `libuild` bin plus `@b9g/libuild/test`. Importing `@b9g/libuild` now fails with `ERR_PACKAGE_PATH_NOT_EXPORTED`.
+- **Main-entry detection no longer guesses.** With nothing declared (no `exports["."]`, `main`, `module`, or `index` entry, multiple entries, no name match), the build now emits NO root export instead of alphabetically electing one - the old fallback would have silently made `.` point at `cli.js` here. Single-entry and name-matching packages are unaffected.
+
 ### Added
 - **Concurrent tests with correct snapshot attribution, via AsyncContext.** Run-time current-test tracking is now an `AsyncContext.Variable` (`@b9g/async-context`, the TC39-shaped shim over `AsyncLocalStorage`) instead of a module global, so interleaved test bodies each see their own name: `toMatchSnapshot()` inside `test.concurrent` (bun) or options-form `test(name, {concurrency}, fn)` (node) files each snapshot under its own key - previously concurrent bodies would have silently misattributed each other's snapshots through the shared global, and `.concurrent` bodies weren't name-tracked or registration-counted at all. `test.concurrent` is also shimmed where the runtime lacks it (node backend, browser runner) as plain registration - sequential execution is a conforming implementation of concurrent semantics - so bun-authored concurrent suites run unchanged everywhere. Adds a `@b9g/async-context` dependency; `node:async_hooks` joins the browser-bundle stubs. `.each` is now implemented once by libuild for every runtime rather than delegated: rows register through the wrapped block, so each row gets its formatted name, a tracked body, and an accurate count identically on bun and node (delegating left rows sharing one snapshot bucket numbered in execution order - fine sequentially, order-dependent under concurrency). Sub-methods are wrapped recursively, so chains (`test.concurrent.each`, `test.skip.each`) survive and runtime-specific variants (`skipIf`, `todoIf`) get tracking for free.
